Move partner-exposed code into an 'Api' namespace
