+++ eng/targets/Imports.targets
@@ -245,8 +245,8 @@
     <ItemGroup>
       <_InternalsVisibleToAttribute Include="System.Runtime.CompilerServices.RestrictedInternalsVisibleToAttribute">
         <_Parameter1>%(RestrictedInternalsVisibleTo.Identity)</_Parameter1>
-        <_Parameter2 Condition="'%(RestrictedInternalsVisibleTo.Partner)' != ''">Microsoft.CodeAnalysis.ExternalAccess.%(RestrictedInternalsVisibleTo.Partner)</_Parameter2>
-        <_Parameter2 Condition="'%(RestrictedInternalsVisibleTo.Partner)' == ''">Microsoft.CodeAnalysis.ExternalAccess.UnknownPartner</_Parameter2>
+        <_Parameter2 Condition="'%(RestrictedInternalsVisibleTo.Partner)' != ''">Microsoft.CodeAnalysis.ExternalAccess.%(RestrictedInternalsVisibleTo.Partner).Api</_Parameter2>
+        <_Parameter2 Condition="'%(RestrictedInternalsVisibleTo.Partner)' == ''">Microsoft.CodeAnalysis.ExternalAccess.UnknownPartner.Api</_Parameter2>
       </_InternalsVisibleToAttribute>
     </ItemGroup>
   </Target>
