+++ src/VisualStudio/Core/Def/ExternalAccess/LegacyCodeAnalysis/Api/ILegacyCodeAnalysisVisualStudioDiagnosticListSuppressionStateServiceAccessor.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-namespace Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis
+namespace Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis.Api
 {
     internal interface ILegacyCodeAnalysisVisualStudioDiagnosticListSuppressionStateServiceAccessor
     {
