+++ src/VisualStudio/Core/Def/ExternalAccess/LegacyCodeAnalysis/Api/ILegacyCodeAnalysisVisualStudioSuppressionFixServiceAccessor.cs
@@ -2,7 +2,7 @@
 
 using Microsoft.VisualStudio.Shell.Interop;
 
-namespace Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis
+namespace Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis.Api
 {
     internal interface ILegacyCodeAnalysisVisualStudioSuppressionFixServiceAccessor
     {
