+++ src/VisualStudio/Core/Def/ExternalAccess/LegacyCodeAnalysis/LegacyCodeAnalysisVisualStudioDiagnosticListSuppressionStateServiceAccessor.cs
@@ -2,10 +2,11 @@
 
 using System;
 using System.Composition;
+using Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis.Api;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.VisualStudio.LanguageServices.Implementation.TableDataSource;
 
-namespace Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis.Internal
+namespace Microsoft.CodeAnalysis.ExternalAccess.LegacyCodeAnalysis
 {
     [Export(typeof(ILegacyCodeAnalysisVisualStudioDiagnosticListSuppressionStateServiceAccessor))]
     [Shared]
